# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Events-API is service for storing Kubernetes and Containerum events in MongoDB.",
    "title": "Events-API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/events/containerum/system": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get system events.",
        "operationId": "GetSystemEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Add system event.",
        "operationId": "AddSystemEvent",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "event added"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/events/containerum/users": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get users events.",
        "operationId": "GetUsersEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Add user event.",
        "operationId": "AddUserEvent",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "event added"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/events/namespaces/{namespace}/pods": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get namespace pods events.",
        "operationId": "GetNamespacePodsEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/events/namespaces/{namespace}/pods/{pod}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get pod events.",
        "operationId": "GetPodEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "pod",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/events/namespaces/{namespace}/pvc": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get namespace PVCs events.",
        "operationId": "GetNamespacePVCsEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/events/namespaces/{namespace}/pvc/{pvc}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get PVC events.",
        "operationId": "GetPVCEventsList",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "pvc",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "200": {
            "description": "events list",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "Event": {
      "description": "Event -- Containerum event",
      "type": "object",
      "properties": {
        "details": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "event_kind": {
          "$ref": "#/definitions/EventKind"
        },
        "event_name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "event_time": {
          "type": "string",
          "x-go-name": "Time"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "resource_name": {
          "type": "string",
          "x-go-name": "ResourceName"
        },
        "resource_namespace": {
          "type": "string",
          "x-go-name": "ResourceNamespace"
        },
        "resource_type": {
          "$ref": "#/definitions/ResourceType"
        },
        "resource_uid": {
          "type": "string",
          "x-go-name": "ResourceUID"
        }
      },
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "EventKind": {
      "type": "string",
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "EventsList": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          },
          "x-go-name": "Events"
        }
      },
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ResourceType": {
      "type": "string",
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "allResourcesChangesEventsDefaultBody": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "description": "ErrID -- represents unique error ID",
          "type": "object",
          "properties": {
            "kind": {
              "description": "ErrKind -- represents kind of error",
              "type": "integer",
              "format": "uint64",
              "x-go-package": "github.com/containerum/cherry"
            },
            "sid": {
              "description": "ErrSID -- represents service ID of error",
              "type": "string",
              "x-go-package": "github.com/containerum/cherry"
            }
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-gen-location": "operations",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}